Major refactoring, separated statements
